Remove the check for valid from the Requester to the Responder.
This allow to have implementation of the Requester that still send
message even outside of a valid lease (not recommended, but it could
be useful in specific case like full cluster down).

Lease TTL and Tickets have been converted to 32bits + the time
unit is now millisecond.

+++ src/main/java/io/reactivesocket/Frame.java
@@ -350,13 +350,13 @@ public class Frame implements Payload
 
     public static class Lease
     {
-        public static long ttl(final Frame frame)
+        public static int ttl(final Frame frame)
         {
             ensureFrameType(FrameType.LEASE, frame);
             return LeaseFrameFlyweight.ttl(frame.directBuffer, 0);
         }
 
-        public static long numberOfRequests(final Frame frame)
+        public static int numberOfRequests(final Frame frame)
         {
             ensureFrameType(FrameType.LEASE, frame);
             return LeaseFrameFlyweight.numRequests(frame.directBuffer, 0);
